requirements.txt
requests>=2.25.0
python-dotenv>=0.19.0
scholarly>=1.7.0
